added 'BreakDown' and 'Octavify' functions
'GetChords' extracts a list of chords from the input file (in Lead Sheet Mini-Language format), and also builds a list of bars, where each list element (bar) is itself a list of chords in that bar.
'BreakDown' converts the root notation into Jython Music notation (e.g. BF instead of Bb, AS instead of A#) and it works out the durations. It outputs two lists: one of the root names (as strings) and one of the corresponding durations (as decimals where 1.0 is one whole bar).
'Octavify' converts a list of root notes into a list of MIDI note values.